Subdivided the character cards text section into its own component
